delete duplication in arrayItem
