+++ ReservacionDeVuelos/src/consts/flights.js
@@ -1,0 +1,12 @@
+const arrayItems = [
+  {label: 'Belgrade, Serbia', value: {subtitle: 'Serbia', title: 'BEG'}},
+  {label: 'Napoles, Italia', value: {subtitle: 'Italy', title: 'NAP'}},
+  {label: 'Oaxaca, México', value: {subtitle: 'México', title: 'OAX'}},
+  {label: 'Berlin, Germany', value: {subtitle: 'Germany', title: 'BER'}},
+  {
+    label: 'Saskatchewan, Canada',
+    value: {subtitle: 'Saskatchewan', title: 'SAS'},
+  },
+];
+
+export default arrayItems;
