+++ ReservacionDeVuelos/src/screens/FromScreen/index.js
@@ -7,17 +7,7 @@ import BackButton from '../../components/atoms/BackButton';
 import BottomButton from '../../components/atoms/BottomButton';
 import functions from './functions';
 import globalstyles from '../../consts/globalstyles';
-
-const arrayItems = [
-  {label: 'Belgrade, Serbia', value: {subtitle: 'Serbia', title: 'BEG'}},
-  {label: 'Napoles, Italia', value: {subtitle: 'Italy', title: 'NAP'}},
-  {label: 'Oaxaca, México', value: {subtitle: 'México', title: 'OAX'}},
-  {label: 'Berlin, Germany', value: {subtitle: 'Germany', title: 'BER'}},
-  {
-    label: 'Saskatchewan, Canada',
-    value: {subtitle: 'Saskatchewan', title: 'SAS'},
-  },
-];
+import arrayItems from '../../consts/flights';
 
 export default function FromScreen({navigation}) {
   const [item, setItem] = useState({});
