+++ ReservacionDeVuelos/src/screens/ToScreen/index.js
@@ -9,17 +9,7 @@ import BackButton from '../../components/atoms/BackButton';
 import BottomButton from '../../components/atoms/BottomButton';
 import globalstyles from '../../consts/globalstyles';
 import texts from '../../consts/text';
-
-const arrayItems = [
-  {label: 'Belgrade, Serbia', value: {subtitle: 'Serbia', title: 'BEG'}},
-  {label: 'Napoles, Italia', value: {subtitle: 'Italy', title: 'NAP'}},
-  {label: 'Oaxaca, México', value: {subtitle: 'México', title: 'OAX'}},
-  {label: 'Berlin, Germany', value: {subtitle: 'Germany', title: 'BER'}},
-  {
-    label: 'Saskatchewan, Canada',
-    value: {subtitle: 'Saskatchewan', title: 'SAS'},
-  },
-];
+import arrayItems from '../../consts/flights';
 
 export default function ToScreen({navigation, route}) {
   const [to, setTo] = useState({});
